+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.20.1 - 2.1.1]
+
+### Added
+- Compatibility with Bean's Backpacks Winged Backpack
+
+### Changed
+- Refactored mod compatibility code to compat package.
+
 ## [1.20.1 - 2.1.0] June 7th, 2023
 
 ### Changed